pyproject.toml
[project]
name = "demo-python-modern"
version = "1.0.0"
description = "Demo Python app with vulnerable dependencies for GHAS testing"
requires-python = ">=3.8"
dependencies = [
    "django==2.2.0",
    "flask==0.12.2",
    "requests==2.6.0",
    "pyyaml==5.1",
    "jinja2==2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
